// src/lib/models/Continent.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum Continent {
    AFRIQUE = 'Afrique',
    AM_RIQUE_DU_NORD = 'Amérique du Nord',
    AM_RIQUE_DU_SUD = 'Amérique du Sud',
    ASIE = 'Asie',
    EUROPE = 'Europe',
    OC_ANIE = 'Océanie',
    ANTARCTIQUE = 'Antarctique',
}