// src/lib/models/ClimateZone.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum ClimateZone {
    TROPICAL = 'tropical',
    SUBTROPICAL = 'subtropical',
    TEMPERATE = 'temperate',
    CONTINENTAL = 'continental',
    ARID = 'arid',
    SEMI_ARID = 'semi-arid',
    MEDITERRANEAN = 'mediterranean',
}